Keygen
This is a keygen I made to reverse engineer a Malware in Reverse engineering class. It grabbed the hex of a password from a location I found using some other software and put it together. Needs to be edited depending on the target, but this is the main concept. Once I would get the key, I would put it the decryption algorithm (again varies based on the target). And the PNG file shows the results. Again, it will look different for each target, but this is a proof of concept.